developer: adjust dir structure and refactor import paths
